Refactor monolithic main.go into internal packages (cli, worker, consoleui, session)
